health checks: factor out get_required_version
