+++ roles/openshift_health_checker/test/ovs_version_test.py
@@ -1,26 +1,7 @@
 import pytest
 
-from openshift_checks.ovs_version import OvsVersion, OpenShiftCheckException
-
-
-def test_openshift_version_not_supported():
-    def execute_module(*_):
-        return {}
-
-    openshift_release = '111.7.0'
-
-    task_vars = dict(
-        openshift=dict(common=dict()),
-        openshift_release=openshift_release,
-        openshift_image_tag='v' + openshift_release,
-        openshift_deployment_type='origin',
-        openshift_service_type='origin'
-    )
-
-    with pytest.raises(OpenShiftCheckException) as excinfo:
-        OvsVersion(execute_module, task_vars).run()
-
-    assert "no recommended version of Open vSwitch" in str(excinfo.value)
+from openshift_checks.ovs_version import OvsVersion
+from openshift_checks import OpenShiftCheckException
 
 
 def test_invalid_openshift_release_format():
